refactor: 유저 행동 추적 기반 컴포넌트 및 훅 폴더 변경
