+++ frontend/techpick/src/app/(signed)/layout.tsx
@@ -4,9 +4,9 @@ import { getTagList } from '@/apis/tag/getTagList';
 import { FeedbackToolbar } from '@/components/FeedbackToolbar';
 import { FolderAndPickDndContextProvider } from '@/components/FolderAndPickDndContextProvider';
 import { MobileNavigationBar } from '@/components/MobileNavigationBar';
-import { ScreenLogger } from '@/components/ScreenLogger';
 import ShortcutKey from '@/components/ShortcutKey';
 import { SideNavigationBar } from '@/components/SideNavigationBar/SideNavigationBar';
+import { ScreenLogger } from '@/libs/@eventlog/ScreenLogger';
 import { getQueryClient } from '@/libs/@react-query/getQueryClient';
 import { folderKeys } from '@/queries/folderKeys';
 import { tagKeys } from '@/queries/tagKeys';
