+++ frontend/techpick/src/libs/@eventlog/ScreenLogger.tsx
@@ -1,10 +1,10 @@
 'use server';
 
-import { mixpanelServer } from '@/libs/mixpanel-server';
 import { getUserIdForServer } from '@/utils/getUserIdForServer';
 import { headers } from 'next/headers';
 import { userAgent } from 'next/server';
 import type { PropsWithChildren } from 'react';
+import { mixpanelServer } from './mixpanel-server';
 
 /**
  * @description 특정 페이지에 방문했는지 확인하는 서버 컴포넌트입니다.
