+++ frontend/techpick/src/libs/@eventlog/useEventLogger.ts
@@ -1,7 +1,7 @@
 'use client';
 
-import mixpanel from '@/libs/mixpanel-client';
 import { useCallback } from 'react';
+import mixpanel from './mixpanel-client';
 
 /**
  * @description 특정 액션에 로그를 추가하는 훅입니다.
